3i-atlas/requirements.txt
# 3I Atlas Analysis Suite Requirements
# =====================================

# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
opencv-python>=4.8.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
PyYAML>=6.0

# Image processing
Pillow>=8.3.0

# Astronomical analysis (optional but recommended)
# astropy>=4.0.0  # Uncomment for advanced astronomical calculations

# Machine learning and statistics
scikit-image>=0.18.0

# Development and testing
pytest>=6.2.0

# Note: This analysis suite is designed to work with standard scientific Python libraries.
# Some advanced features may require additional packages, which will be noted in the documentation.